requirements.txt
pandas
numpy
matplotlib
scikit-learn
statsmodels
python-dotenv
hdbcli
openai
fastapi
uvicorn
requests
pytest
joblib
